# resources/swagger.yaml
openapi: 3.0.1
info:
  title: Symfony
  description: 'Это спецификация API для тестового приложения symfony.local'
  contact:
    email: mixno09@mail.ru
  version: 1.0.0
servers:
  - url: http://symfony.local/api
paths:
  /products:
    post:
      tags:
        - products
      summary: Добавление нового продукта в БД
      requestBody:
        description: 'Загружаемое изображение должно быть размером 400x400 до 800x800'
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                title:
                  type: string
                  minLength: 5
                  maxLength: 255
                slug:
                  type: string
                  pattern: '^(?!-)([a-z0-9]|(?<!-)-)+(?<!-)$'
                description:
                  type: string
                  minLength: 5
                image:
                  type: string
                  format: binary
              required:
                - id
                - title
                - slug
                - description
                - image
            encoding:
              image:
                contentType: image/jpeg
      responses:
        201:
          description: Продукт создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                required:
                  - url
        422:
          description: Ошибка валидации
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
                required:
                  - errors
  /products/{id}:
    get:
      tags:
        - products
      summary: Получение продукта по ID
      parameters:
      - name: id
        in: path
        schema:
          type: string
          format: uuid
        required: true
      responses:
        200:
          description: Продукт
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  slug:
                    type: string
                  description:
                    type: string
                  image:
                    type: string
                    format: uri
                required:
                  - id
                  - title
                  - slug
                  - description
                  - image
        404:
          description: Продукт не найден